+++ src/libs/relay/conduit_relay_mpi_io_blueprint.hpp
@@ -171,6 +171,24 @@ void CONDUIT_RELAY_API read_mesh(const std::string &root_file_path,
                                  conduit::Node &mesh,
                                  MPI_Comm comm);
 
+//-----------------------------------------------------------------------------
+void CONDUIT_RELAY_API load_mesh(const std::string &root_file_path,
+                                 conduit::Node &mesh,
+                                 MPI_Comm comm);
+
+
+//-----------------------------------------------------------------------------
+///
+/// opts:
+///      mesh_name: "{name}"
+///          provide explicit mesh name, for cases where bp data includes
+///           more than one mesh.
+//-----------------------------------------------------------------------------
+void CONDUIT_RELAY_API load_mesh(const std::string &root_file_path,
+                                 const conduit::Node &opts,
+                                 conduit::Node &mesh,
+                                 MPI_Comm comm);
+
 }
 //-----------------------------------------------------------------------------
 // -- end conduit::relay::mpi::io::blueprint --